+++ CHANGELOG.md
@@ -4,9 +4,12 @@
 
 - 新增 `f-input` 组件
 - 废除 `f-dialog` 组件 `close-icon` 属性
-- 新增 `f-dialog` 组件 `close-icon` 插槽
+- 新增 `f-dialog` 组件 `closeIcon` 插槽
 - 废除 `f-drawer` 组件 `close-icon` 属性
-- 新增 `f-drawer` 组件 `close-icon` 插槽
+- 新增 `f-drawer` 组件 `closeIcon` 插槽
+- 废除 `f-alert` 组件 `icon` 属性
+- 新增 `f-alert` 组件 `prefixIcon` 和 `closeIcon` 插槽
+- 新增 `f-close-btn` 组件 `color` 属性
 
 ## 0.7.0-alpha.9 (2022-10-19)